Implement red-black tree sets
Woo-wee.  This is a bit nasty without pattern matching.  The removal
algorithm is also a bit tough.

This is basically Kahrs's version of red-black trees.  Interestingly,
his implementation differs from Okasaki's in the insert and balance
functions.  Where Okasaki runs balance every time, Karhs only runs
balance when going through a black node.  They also have different
opinions on what color node balance should return when no balancing is
required.

Both Kahrs and Okasaki seem to try to avoid consing by decomposing
trees into their components and then they pass those components into
each function.  That way if you want to swap out one of those
components you just change the value passed to the function.  That's
works well for them, but we have two different types of nodes, and
most functions want to treat them the same way.  E.g. balance doesn't
care whether we're visiting an uneql-tree-node or a tree-node.  Rather
than pass in the expected type of the node (so balance can make a new
node of the right type), I thought it would be easier to just pass in
the node holding the value and any left/right substitutions as
arguments.  Then the receiver uses tree-node-with-changes to swap out
children or color as required.  If we want to change the children the
receiver sees, we'd need to cons up a new tree node... or do we?  I
developed a convention of passing in replacement left and right
children into functions along with the tree.  That way we don't need
to cons up a new node just to pass it to some other function.  I think
this captures the benefits of Kahrs's and Okasaki's approach without
too much extra confusion.  I hope.

As with unbalanced sets, handling :unequal things added a lot of extra
complexity.  In addition to causing issues (like above), it also
motivated small changes to the algorithms.  E.g. removing an item may
not change the tree's balance at all!  Re-balancing needs to be
conditional.

Phew.  It seems to work, thankfully.