Hijack the copier for immutable structures
This isn't how I'd like it to work long-term.  I'd prefer to have a
proper metaclass-like setup where you declare the metaclass of the
struct and that influences how things are generated.  E.g. a
IMMUTABLE-STRUCTURE metaclass might generate this sort of copier
rather than the default.  But... this little diversion has gone on
long enough.  Its time to actually leverage the auto-generated copier.