Partition comparisons by classes
Subclassing in the old scheme was hazardous.  If a subclass didn't
define comparison methods for all of its superclasses (both direct and
transitive!) then you could get a bogus result.  This could quite
easily result in intransitivity for the comparison ordering.

Imagine a base class with no slots.  It defines a comparison method
that always returns :equal.  A foolish subclass comes along and adds
slots but forgets to add a method for comparing against base class
instances.  Now you can have two subclass instances which are not
:equal to each other, but are both :equal to an instance of the base
class.  Whoops.

This is actually a huge problem.  Once you allow people to subclass
your class, you aren't allowed to change the superclasses of your
class any more.  This makes classes extremely fragile, and I'm not
comfortable with that!

... this does have some unfortunate consequences which we will deal
with shortly.