Use eq instead of eql for comparison functions
EXCEPT in...
- the compare function itself
- compare-objects-using-slots
- compare-reals

In these contexts we don't know whether the arguments are numbers or
not.

After this, here's a side-by-side perf comparison of fset and us when
we are compiled with 3 speed, 2 space, and 0 for everything else.

CL-USER> (defvar *objects* (coerce (pfds.shcl.io/tests/compare::make-test-objects) 'vector))
*OBJECTS*
CL-USER> (progn
           (sb-ext:gc :full t)
           (time (dotimes (i 100000)
                   (pfds.shcl.io/tests/compare::sorted #'pfds.shcl.io/compare::compare *objects*))))
Evaluation took:
  2.791 seconds of real time
  2.792112 seconds of total run time (2.780097 user, 0.012015 system)
  [ Run times consist of 0.009 seconds GC time, and 2.784 seconds non-GC time. ]
  100.04% CPU
  10,050,418,426 processor cycles
  252,890,656 bytes consed

NIL
CL-USER> (progn
           (sb-ext:gc :full t)
           (time (dotimes (i 100000)
                   (pfds.shcl.io/tests/compare::sorted #'fset:compare *objects*))))
Evaluation took:
  3.267 seconds of real time
  3.267703 seconds of total run time (3.262710 user, 0.004993 system)
  [ Run times consist of 0.016 seconds GC time, and 3.252 seconds non-GC time. ]
  100.03% CPU
  11,760,222,546 processor cycles
  433,584,256 bytes consed

NIL

These results are reproducible... on my system at least.

Now, to be fair, fset doesn't appear to compile its compare method
with optimization on.  So, its a bit unfair for us to have
optimizations.  Meh?