Fix an algorithmic bug in weight balanced trees
While investigating a failing unit test I uncovered a bug in the
paper I was using as my reference for weight balanced trees.  As I
continued to investigate I found that I wasn't the only one who had
noticed the issue.  Haskell's implementation of weight balanced trees
pointed me toward a paper that confirmed the bug was real and provided
the solution I adopted here.

FSet seems to take a different approach.  They implement the algorithm
exactly as Adams described in his original paper... almost.  They run
their balancer function on the subtrees when performing a rotation.
In my (brief) testing, that approach seems to work, but I have no
reason to believe it won't result in a flurry of unnecessary balancing
operations that ripple through the tree.  I'm going to stick with the
approach that seems to have been proven safe!